+++ tests/tinytest/test_funs_other.R
@@ -81,3 +81,59 @@ expect_equal(
   c(1, 2)
 )
 
+# length
+
+test <- pl$DataFrame(x = c("a", "a", "a", "b", "b"), y = c(1:4, NA))
+
+expect_equal(
+  test |>
+    mutate(foo = length(y)) |>
+    pull(foo),
+  rep(5, 5)
+)
+
+expect_equal(
+  test |>
+    mutate(foo = length(y), .by = x) |>
+    pull(foo),
+  c(3, 3, 3, 2, 2)
+)
+
+expect_equal(
+  test |>
+    mutate(foo = length(y), .by = c(x, y)) |>
+    pull(foo),
+  rep(1, 5)
+)
+
+
+# unique
+
+test <- pl$DataFrame(x = c("a", "a", "a", "b", "b"), y = c(2, 2, 3, 4, NA))
+
+expect_error(
+  test |>
+    mutate(foo = unique(y)),
+  "lengths don't match"
+)
+
+expect_equal(
+  test |>
+    mutate(foo = length(unique(y))) |>
+    pull(foo),
+  rep(4, 5)
+)
+
+expect_equal(
+  test |>
+    mutate(foo = length(unique(y)), .by = x) |>
+    pull(foo),
+  rep(2, 5)
+)
+
+expect_equal(
+  test |>
+    mutate(foo = length(unique(y)), .by = c(x, y)) |>
+    pull(foo),
+  rep(1, 5)
+)